chore: moved tests to utils/__tests__  folder
