+++ src/fragmentation/utils/__tests__/fragment.test.js
@@ -1,6 +1,6 @@
 import OCL from 'openchemlib';
 
-import { fragment } from '../fragment.js';
+import { fragment } from '../../fragment.js';
 
 const { Molecule } = OCL;
